Move DeepSSZEqual assertions to shared/sszutil
