+++ beacon-chain/operations/slashings/BUILD.bazel
@@ -44,6 +44,7 @@ go_test(
         "//beacon-chain/state:go_default_library",
         "//shared/bls:go_default_library",
         "//shared/params:go_default_library",
+        "//shared/sszutil:go_default_library",
         "//shared/testutil:go_default_library",
         "//shared/testutil/assert:go_default_library",
         "//shared/testutil/require:go_default_library",
