+++ shared/testutil/assert/assertions.go
@@ -25,16 +25,6 @@ func DeepNotEqual(tb assertions.AssertionTestingTB, expected, actual interface{}
 	assertions.DeepNotEqual(tb.Errorf, expected, actual, msg...)
 }
 
-// DeepSSZEqual compares values using sszutil.DeepEqual.
-func DeepSSZEqual(tb assertions.AssertionTestingTB, expected, actual interface{}, msg ...interface{}) {
-	assertions.DeepSSZEqual(tb.Errorf, expected, actual, msg...)
-}
-
-// DeepNotSSZEqual compares values using sszutil.DeepEqual.
-func DeepNotSSZEqual(tb assertions.AssertionTestingTB, expected, actual interface{}, msg ...interface{}) {
-	assertions.DeepNotSSZEqual(tb.Errorf, expected, actual, msg...)
-}
-
 // NoError asserts that error is nil.
 func NoError(tb assertions.AssertionTestingTB, err error, msg ...interface{}) {
 	assertions.NoError(tb.Errorf, err, msg...)
