+++ shared/testutil/assertions/BUILD.bazel
@@ -7,7 +7,6 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/shared/testutil/assertions",
     visibility = ["//visibility:public"],
     deps = [
-        "//shared/sszutil:go_default_library",
         "@com_github_d4l3k_messagediff//:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
