+++ shared/testutil/assertions/assertions.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/d4l3k/messagediff"
 	"github.com/gogo/protobuf/proto"
-	"github.com/prysmaticlabs/prysm/shared/sszutil"
 	"github.com/sirupsen/logrus/hooks/test"
 )
 
@@ -52,25 +51,6 @@ func DeepEqual(loggerFn assertionLoggerFn, expected, actual interface{}, msg ...
 // DeepNotEqual compares values using DeepEqual.
 func DeepNotEqual(loggerFn assertionLoggerFn, expected, actual interface{}, msg ...interface{}) {
 	if isDeepEqual(expected, actual) {
-		errMsg := parseMsg("Values are equal", msg...)
-		_, file, line, _ := runtime.Caller(2)
-		loggerFn("%s:%d %s, want: %#v, got: %#v", filepath.Base(file), line, errMsg, expected, actual)
-	}
-}
-
-// DeepSSZEqual compares values using sszutil.DeepEqual.
-func DeepSSZEqual(loggerFn assertionLoggerFn, expected, actual interface{}, msg ...interface{}) {
-	if !sszutil.DeepEqual(expected, actual) {
-		errMsg := parseMsg("Values are not equal", msg...)
-		_, file, line, _ := runtime.Caller(2)
-		diff, _ := messagediff.PrettyDiff(expected, actual)
-		loggerFn("%s:%d %s, want: %#v, got: %#v, diff: %s", filepath.Base(file), line, errMsg, expected, actual, diff)
-	}
-}
-
-// DeepNotSSZEqual compares values using sszutil.DeepEqual.
-func DeepNotSSZEqual(loggerFn assertionLoggerFn, expected, actual interface{}, msg ...interface{}) {
-	if sszutil.DeepEqual(expected, actual) {
 		errMsg := parseMsg("Values are equal", msg...)
 		_, file, line, _ := runtime.Caller(2)
 		loggerFn("%s:%d %s, want: %#v, got: %#v", filepath.Base(file), line, errMsg, expected, actual)
