+++ shared/testutil/require/requires.go
@@ -25,16 +25,6 @@ func DeepNotEqual(tb assertions.AssertionTestingTB, expected, actual interface{}
 	assertions.DeepNotEqual(tb.Fatalf, expected, actual, msg...)
 }
 
-// DeepSSZEqual compares values using DeepEqual.
-func DeepSSZEqual(tb assertions.AssertionTestingTB, expected, actual interface{}, msg ...interface{}) {
-	assertions.DeepSSZEqual(tb.Fatalf, expected, actual, msg...)
-}
-
-// DeepNotSSZEqual compares values using DeepEqual.
-func DeepNotSSZEqual(tb assertions.AssertionTestingTB, expected, actual interface{}, msg ...interface{}) {
-	assertions.DeepNotSSZEqual(tb.Fatalf, expected, actual, msg...)
-}
-
 // NoError asserts that error is nil.
 func NoError(tb assertions.AssertionTestingTB, err error, msg ...interface{}) {
 	assertions.NoError(tb.Fatalf, err, msg...)
